Fixed issue with one-hot encoder when encoding data with more than one feature.
Previously, when encoding a dataset with multiple features, it was encoded as if each sample was a feature. This resulted in every one-hot vector having only a single 1. Now, each feature is encoded as a one-hot vector so a single will have as many 1s as there are features per sample.