Several changes to Linear and Dense layers back() method.
upstream gradients are now the transpose of what they were before to eliminate unneeded transposing.

Fix for gradient computations in the Dense layer.

In the Dense layers back() method, there was a computation which was being done twice unnecessarily. This has been changed so that the computation is only done a single time.